misc changes
prepare for chunk management

+++ src/block.hpp
@@ -2,10 +2,11 @@
 #define BLOCK_HPP
 
 #include <glm/vec3.hpp>
+#include <sys/types.h>
 
 namespace kwnc
 {
-enum class Block : int { AIR, DIRT, STONE };
+enum class Block : u_int16_t { AIR, DIRT, STONE };
 
 struct Block_Vertex {
         glm::vec3 position; // Represents chunk coordinates
